Add files via upload
We can employ density-based approaches, which represent clusters as dense regions split by sparse regions, to locate clusters with observable non-spherical forms.
Density-Based Spatial Clustering of Applications with Noise (DBSCAN) is a technique that aims to aggregate spatial regions with data point densities that are higher than a set threshold.
In order to use DBSCAN, a specified distance known as eps (for epsilon), which is the distance between points to be treated as a cluster, and min_ samples, which is the minimum amount of data points necessary to build a dense cluster region, are required. These variables help regulate the algorithm's tolerance for noise and the characteristics of the distribution that should be treated as a cluster.